backend/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==1.0.0
requests==2.31.0
python-chess==1.2.0
SQLAlchemy==1.4.46
